#2 Fix date bug
In the 'Ends local' column the date of the task's start was shown. This is fixed and now the correct date is displayed. This issue was present for all end dates but only was noticeable if the end date in the local time zone was on another day than the start.