#7 Add delay of 100ms before start calculation
To prevent that local times are calculated on old data a delay was added before calculation starts.
Furthermore the delay is reset if another change which is originated in one of the default table columns (i.e. not the columns added by the extension).
This leads to about 1/4 of the number of executions of the recalculation, which also should improve performance